reword
@eddelbuettel, thanks for your feedback, here's an attempt to be more general, acknowledge multiple containers in the repo, direct the reader to the wiki, and give them a line they can use to get started immediately. What do you think?

I'm not sure of the best practice for starting a docker container, I'd welcome advice on suitable arguments to use for the casual tire-kicking first-timer rocker user. I saw `--rm -t -i ` in a few places in other issue threads on this repo, so that seemed like a good place to start.