split 'overview' and 'getting started'
Yes, I hesitated about merging the two sections, and your comment confirms that was a poor choice. Here they are separate again. In this PR, the 'getting started' section is distinct but still visible to the casual browser coming from twitter, etc., without a minimum of scrolling. Hope that helps!