Improve user options
If an alternative username is requested, the script now
renames the existing user instead of deleting `docker` user
and creating a new user.  The previous behavoir could cause
trouble if the container was already run as the docker user.

This complements the previous push which allows the container
to be run using the default argument (supervisord) by a user
other than root (though custom users cannot be configured in
this case without the use of sudo, as userconf.sh requires
root privileges).

A new user is added only if a new UID is requested for the user
(e.g. when sharing volumes and on the host who's UID differs from
1000.)